requirements.txt
asgiref==3.8.1
attrs==24.1.0
autobahn==24.4.2
Automat==22.10.0
blinker==1.8.2
Brotli==1.1.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.0
cssselect2==0.7.0
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.4.0
django-crontab==0.7.1
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.53.1
html5lib==1.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.0.8
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyphen==0.16.0
python-dotenv==1.0.1
redis==5.0.8
reportlab==4.2.2
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tinycss2==1.3.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
zope.interface==6.4.post2
zopfli==0.2.3
